String: performance improvement converting uint64_t to string (4x)
Based on the trick given by A. Alexandrescu on Facebook Engineering
blog, this patch replace the old mk_str_itop() with the faster
algorithm proposed.

More details about the implementation here:

 https://www.facebook.com/notes/facebook-engineering/three-optimization-tips-for-c/10151361643253920

+++ src/include/mk_plugin.h
@@ -208,7 +208,7 @@ struct plugin_api
     char *(*pointer_to_buf) (mk_ptr_t);
 
     /* string functions */
-    int   (*str_itop) (int, mk_ptr_t *);
+    int   (*str_itop) (uint64_t, mk_ptr_t *);
     int   (*str_search) (const char *, const char *, int);
     int   (*str_search_n) (const char *, const char *, int, int);
     char *(*str_build) (char **, unsigned long *, const char *, ...) PRINTF_WARNINGS(3,4);
